Modified printHelp method
Modified the printHelp() method for practice with JSE 15 comment blocks.

Used int version to collect the version used and pass it to the method, but JSE < 14 will not compile a comment block, so I left it with commented todos.

No significant functionality change

Added CSV results of the program running operations.